+++ flatlaf-testing/src/main/java/com/formdev/flatlaf/testing/FlatComponentsTest.java
@@ -93,21 +93,25 @@ public class FlatComponentsTest
 		JLabel spinnerLabel = new JLabel();
 		JSpinner spinner1 = new JSpinner();
 		JSpinner spinner2 = new JSpinner();
+		JComboBox<String> comboBox7 = new JComboBox<>();
 		JLabel textFieldLabel = new JLabel();
 		JTextField textField1 = new JTextField();
 		JTextField textField2 = new JTextField();
 		JTextField textField3 = new JTextField();
 		JTextField textField4 = new JTextField();
+		JTextField textField6 = new JTextField();
 		JLabel formattedTextFieldLabel = new JLabel();
 		JFormattedTextField formattedTextField1 = new JFormattedTextField();
 		JFormattedTextField formattedTextField2 = new JFormattedTextField();
 		JFormattedTextField formattedTextField3 = new JFormattedTextField();
 		JFormattedTextField formattedTextField4 = new JFormattedTextField();
+		JFormattedTextField formattedTextField5 = new JFormattedTextField();
 		JLabel passwordFieldLabel = new JLabel();
 		JPasswordField passwordField1 = new JPasswordField();
 		JPasswordField passwordField2 = new JPasswordField();
 		JPasswordField passwordField3 = new JPasswordField();
 		JPasswordField passwordField4 = new JPasswordField();
+		JPasswordField passwordField5 = new JPasswordField();
 		JLabel textAreaLabel = new JLabel();
 		JScrollPane scrollPane1 = new JScrollPane();
 		JTextArea textArea1 = new JTextArea();
@@ -447,6 +451,11 @@ public class FlatComponentsTest
 		spinner2.setEnabled(false);
 		add(spinner2, "cell 2 6,growx");
 
+		//---- comboBox7 ----
+		comboBox7.setEditable(true);
+		comboBox7.putClientProperty("JTextField.placeholderText", "placeholder");
+		add(comboBox7, "cell 5 6,growx");
+
 		//---- textFieldLabel ----
 		textFieldLabel.setText("JTextField:");
 		add(textFieldLabel, "cell 0 7");
@@ -470,6 +479,10 @@ public class FlatComponentsTest
 		textField4.setEnabled(false);
 		textField4.setEditable(false);
 		add(textField4, "cell 4 7,growx");
+
+		//---- textField6 ----
+		textField6.putClientProperty("JTextField.placeholderText", "placeholder");
+		add(textField6, "cell 5 7,growx");
 
 		//---- formattedTextFieldLabel ----
 		formattedTextFieldLabel.setText("JFormattedTextField:");
@@ -495,6 +508,10 @@ public class FlatComponentsTest
 		formattedTextField4.setEditable(false);
 		add(formattedTextField4, "cell 4 8,growx");
 
+		//---- formattedTextField5 ----
+		formattedTextField5.putClientProperty("JTextField.placeholderText", "placeholder");
+		add(formattedTextField5, "cell 5 8,growx");
+
 		//---- passwordFieldLabel ----
 		passwordFieldLabel.setText("JPasswordField:");
 		add(passwordFieldLabel, "cell 0 9");
@@ -518,6 +535,10 @@ public class FlatComponentsTest
 		passwordField4.setEnabled(false);
 		passwordField4.setEditable(false);
 		add(passwordField4, "cell 4 9,growx");
+
+		//---- passwordField5 ----
+		passwordField5.putClientProperty("JTextField.placeholderText", "placeholder");
+		add(passwordField5, "cell 5 9,growx");
 
 		//---- textAreaLabel ----
 		textAreaLabel.setText("JTextArea:");